index_service/pyproject.toml
[tool.poetry]
name = "index-service"
version = "0.1.0"
description = ""
authors = ["ce.liu <ce.liu@avepoint.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.12"
fastapi = "^0.115.6"
llama-index = "^0.12.2"
uvicorn = "^0.32.1"
llama-index-embeddings-ollama = "^0.5.0"
llama-index-llms-ollama = "^0.5.0"
chromadb = "^0.5.23"
llama-index-vector-stores-chroma = "^0.4.1"
openpyxl = "^3.1.5"
docx2txt = "^0.8"
torch = "^2.5.1"
transformers = "4.46.0"
python-pptx = "^1.0.2"
pillow = "^11.0.0"
pypdf = "^5.1.0"
llama-index-graph-stores-neo4j = "^0.4.4"
tenacity = "^9.0.0"
apscheduler = "^3.11.0"
nebula3-python = "^3.8.3"
llama-index-graph-stores-nebula = "^0.4.0"
pytest = "^8.3.4"
pydantic-settings = "^2.7.1"
trulens = "^1.3.5"
trulens-apps-llamaindex = "^1.2.11"
python-multipart = "^0.0.20"
xlrd = "^2.0.1"
graspologic = "3.4.1"
gensim = "^4.3.3"
numpy = "1.26.4"
pandas = "^2.2.3"
trulens-eval = "^1.3.5"
llama-index-llms-azure-openai = "^0.3.0"
python-louvain = "^0.16"
qdrant-client = "^1.13.3"
llama-index-vector-stores-qdrant = "^0.4.3"


[tool.poetry.group.dev.dependencies]
chardet = "^5.2.0"
matplotlib = "^3.10.0"
pygraphviz = "^1.14"

[isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
